+++ src/classes/Application/API/Response/JSONInfoSerializer.php
@@ -1,21 +1,25 @@
 <?php
 /**
- * package API
- * @subpackage Core
+ * @package API
+ * @subpackage Response
  */
 
 declare(strict_types=1);
 
-namespace Application\API;
+namespace Application\API\Response;
+
+use Application\API\APIMethodInterface;
+use Application\API\Traits\JSONResponseTrait;
 
 /**
- * Class used to serialize essential information about an API method
- * for use in API responses.
+ * Helper class that serializes the information on an API method
+ * to an array that is included in JSON responses.
  *
  * @package API
- * @subpackage Core
+ * @subpackage Response
+ * @see JSONResponseTrait::_sendJSONData()
  */
-class APIInfo
+class JSONInfoSerializer
 {
     public const string KEY_REQUEST_MIME = 'requestMime';
     public const string KEY_SELECTED_VERSION = 'selectedVersion';
@@ -25,7 +29,6 @@ class APIInfo
     public const string KEY_DESCRIPTION = 'description';
     public const string KEY_AVAILABLE_VERSIONS = 'availableVersions';
     public const string KEY_DOCUMENTATION_URL = 'documentationURL';
-    public const string KEY_GROUP = 'group';
 
     private APIMethodInterface $method;
 
@@ -41,7 +44,6 @@ class APIInfo
             self::KEY_SELECTED_VERSION => $this->method->getActiveVersion(),
             self::KEY_AVAILABLE_VERSIONS => $this->method->getVersions(),
             self::KEY_DESCRIPTION => $this->method->getDescription(),
-            self::KEY_GROUP => $this->method->getGroup()->getLabel(),
             self::KEY_REQUEST_MIME => $this->method->getRequestMime(),
             self::KEY_RESPONSE_MIME => $this->method->getResponseMime(),
             self::KEY_REQUEST_TIME => $this->method->getRequestTime()?->getISODate(true),